Fix vertical offset bug for equal columns and rows
Odd numbered columns didn't get vertical offset applied when x_hex == y_hex, as drawBase was being passed the value of y_hex and then comparing it to x_hex to determine if an offset should be applied. Replaced the value being passed to drawBase to be zero or one, and then used this as a multiplier to calculate how much of a vertical offset is required.